code/requirements.txt
# Aggregate requirements for the workspace (installs both services)

# Core
python-dotenv==1.0.1
requests>=2.32.0
sentry-sdk==1.45.0

# AI Subtitle service
Flask==2.3.3
Flask-Cors==4.0.0
Werkzeug==2.3.7
boto3>=1.28.0
Unidecode==1.3.8

# Image Creation service
pillow==11.3.0
boto3>=1.34.0
